Customer view refactoring
