+++ app/src/main/java/gachon/termproject/danggeun/Adapter/FirstAdapter.java
@@ -11,7 +11,6 @@ import android.widget.Button;
 import android.widget.EditText;
 import android.widget.ImageView;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AlertDialog;
@@ -24,15 +23,11 @@ import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.firebase.database.annotations.NotNull;
 import com.google.firebase.firestore.FirebaseFirestore;
 
-import java.text.SimpleDateFormat;
 import java.util.ArrayList;
-import java.util.Locale;
 
-import gachon.termproject.danggeun.BreadInfo;
+import gachon.termproject.danggeun.Util.Model.BreadInfo;
 import gachon.termproject.danggeun.Bread_list_manager;
-import gachon.termproject.danggeun.MenuRegister;
 import gachon.termproject.danggeun.R;
-import gachon.termproject.danggeun.StoreRegister;
 import gachon.termproject.danggeun.Util.Firestore;
 
 import static gachon.termproject.danggeun.Util.others.isStorageUrl;
