+++ app/src/main/java/gachon/termproject/danggeun/Customer/ReservationActivity.java
@@ -1,4 +1,4 @@
-package gachon.termproject.danggeun;
+package gachon.termproject.danggeun.Customer;
 
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.AppCompatButton;
@@ -8,6 +8,8 @@ import androidx.recyclerview.widget.RecyclerView;
 import android.os.Bundle;
 
 import java.util.ArrayList;
+
+import gachon.termproject.danggeun.R;
 
 public class ReservationActivity extends AppCompatActivity {
