+++ app/src/main/java/gachon/termproject/danggeun/Util/Model/CartBreadItem.java
@@ -1,4 +1,4 @@
-package gachon.termproject.danggeun;
+package gachon.termproject.danggeun.Util.Model;
 
 public class CartBreadItem {
     String cartBreadName;
