+++ app/src/main/res/layout/activity_reservation.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ReservationActivity">
+    tools:context=".Customer.ReservationActivity">
 
     <androidx.appcompat.widget.AppCompatTextView
         android:id="@+id/store_reservaton"
