+++ app/src/main/res/layout/activity_store.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".StoreActivity">
+    tools:context=".Customer.StoreActivity">
 
     <androidx.appcompat.widget.AppCompatImageView
         android:id="@+id/store_img"
@@ -36,7 +36,7 @@
         android:layout_height="0dp"
         android:layout_gravity="bottom"
         android:layout_marginTop="20dp"
-        app:layout_constraintBottom_toTopOf="@id/btn_reserve"
+        app:layout_constraintBottom_toTopOf="@id/btn_cart"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/store_name" />
@@ -51,20 +51,6 @@
         android:text="장바구니"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toStartOf="@id/btn_reserve"/>
-
-    <androidx.appcompat.widget.AppCompatButton
-        android:background="@color/purple_500"
-        android:textColor="@color/white"
-        android:layout_margin="20dp"
-        android:id="@+id/btn_reserve"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        app:layout_constraintStart_toEndOf="@id/btn_cart"
-        app:layout_constraintBottom_toBottomOf="parent"
-        android:text="예약하기"
         app:layout_constraintEnd_toEndOf="parent"/>
-
-
 
 </androidx.constraintlayout.widget.ConstraintLayout>
