+++ app/src/main/res/layout/bread_store.xml
@@ -7,6 +7,7 @@
 
     <androidx.appcompat.widget.AppCompatImageView
         android:id="@+id/bread_img"
+        app:layout_constraintHorizontal_chainStyle="spread"
         android:layout_width="100dp"
         android:layout_height="100dp"
         tools:src="@drawable/ic_launcher_foreground"
@@ -47,7 +48,16 @@
         app:layout_constraintHorizontal_bias="1.0"
         app:layout_constraintStart_toStartOf="@id/bread_name"
         app:layout_constraintTop_toBottomOf="@id/bread_name"
-        tools:text="9,999 원" />
+        tools:text="9,999" />
+
+    <androidx.appcompat.widget.AppCompatTextView
+        android:id="@+id/tx_bread_end"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="원"
+        app:layout_constraintStart_toEndOf="@id/bread_price"
+        app:layout_constraintTop_toTopOf="@id/bread_price"
+        app:layout_constraintBottom_toBottomOf="@id/bread_price"/>
 
     <androidx.appcompat.widget.AppCompatButton
         android:background="@color/purple_500"
@@ -56,11 +66,11 @@
         android:textColor="@color/white"
         android:textSize="20sp"
         android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        app:layout_constraintStart_toEndOf="@id/bread_name"
+        android:layout_height="0dp"
+        app:layout_constraintStart_toEndOf="@id/tx_bread_end"
         app:layout_constraintTop_toTopOf="@id/bread_name"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintBottom_toBottomOf="@id/bread_price"
+        app:layout_constraintBottom_toBottomOf="@id/tx_bread_end"
         />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
